backend/pyproject.toml
[project]
name = "novelverse-backend"
version = "0.1.0"
description = "NovelVerse API — Vietnamese Chinese web novel reading platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic-settings>=2.0",
    "supabase>=2.0",
    "python-jose[cryptography]>=3.3",
    "bleach>=6.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "opencc-python-reimplemented>=0.1.7",
    "google-generativeai>=0.8.6",
    "upstash-redis>=1.6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.9.0",
    "mypy>=1.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "E702"]
